third_party/WebKit/LayoutTests/fast/dom/custom/class-side-inheritance.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<script>
test(function () {
    var protoA = Object.create(HTMLDivElement.prototype);
    var A = document.register('x-a', {prototype: protoA});
    assert_equals(
        Object.getPrototypeOf(A), HTMLDivElement,
        'generated constructor prototype should be base element constructor ' +
        '(extend built-in element)');

    var protoB = Object.create(protoA);
    var B = document.register('x-b', {prototype: protoB});
    assert_equals(
        Object.getPrototypeOf(B), A,
        'generated constructor prototype should be base element constructor ' +
        '(extend Custom Element)');
}, 'generated constructor prototype property');
</script>
